+++ src/components/whatsAppTemplates/FormTabContentButtons.vue
@@ -380,7 +380,7 @@
         });
       },
       addButton() {
-        if (this.currentButtonType === 'quick_reply') {
+        if (this.buttonsType === 'quick_reply') {
           this.buttons.push({ button_type: 'QUICK_REPLY', text: '' });
         } else if (this.buttons[0].button_type === 'URL') {
           this.buttons.push({ button_type: 'URL' });